+++ app/Mocks/WorkoutsList.ts
@@ -1,0 +1,32 @@
+export const workoutLists = {
+  yoga: [
+    { name: "Sun Salutation", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+    { name: "Tree Pose", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+    { name: "Warrior Pose", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+  ],
+  gym: [
+    { name: "Bench Press", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+    { name: "Squats", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+    { name: "Deadlifts", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+  ],
+  cardio: [
+    { name: "Running", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+    { name: "Cycling", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+    { name: "Jump Rope", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+  ],
+  stretch: [
+    { name: "Hamstring Stretch", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+    { name: "Quad Stretch", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+    { name: "Shoulder Stretch", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+  ],
+  fullbody: [
+    { name: "Burpees", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+    { name: "Mountain Climbers", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+    { name: "Plank", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+  ],
+  legs: [
+    { name: "Lunges", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+    { name: "Leg Press", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+    { name: "Calf Raises", description: 'With Rachael Wisdom', image: "/Exercises/yoga-workout.png" },
+  ],
+};
